Generic.Random.DerivingVia

Documentation

newtype GenericArbitrary weights a

Pick a constructor with a given distribution, and fill its fields with recursive calls to arbitrary.

Example

data X = ...
  deriving Arbitrary via (GenericArbitrary '[2, 3, 5] X)

Picks the first constructor with probability 2/10, the second with probability 3/10, the third with probability 5/10.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitrary.

Since: 1.5.0.0

Constructors

GenericArbitrary

Fields

unGenericArbitrary :: a

newtype GenericArbitraryU a

Pick every constructor with equal probability.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryU.

Since: 1.5.0.0

Constructors

GenericArbitraryU

Fields

unGenericArbitraryU :: a

newtype GenericArbitrarySingle a

arbitrary for types with one constructor. Equivalent to GenericArbitraryU, with a stricter type.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitrarySingle.

Since: 1.5.0.0

Constructors

GenericArbitrarySingle

Fields

unGenericArbitrarySingle :: a

newtype GenericArbitraryRec weights a

Decrease size at every recursive call, but don't do anything different at size 0.

data X = ...
  deriving Arbitrary via (GenericArbitraryRec '[2, 3, 5] X)

N.B.: This replaces the generator for fields of type [t] with listOf' arbitrary instead of listOf arbitrary (i.e., arbitrary for lists).

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryRec.

Since: 1.5.0.0

Constructors

GenericArbitraryRec

Fields

unGenericArbitraryRec :: a

newtype GenericArbitraryG genList weights a

GenericArbitrary with explicit generators.

Example

data X = ...
  deriving Arbitrary via (GenericArbitraryG CustomGens '[2, 3, 5] X)

where, for example, custom generators to override String and Int fields might look as follows:

type CustomGens = CustomString :+ CustomInt

Note on multiple matches

Multiple generators may match a given field: the first will be chosen.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryG.

Since: 1.5.0.0

Constructors

GenericArbitraryG

Fields

unGenericArbitraryG :: a

newtype GenericArbitraryUG genList a

GenericArbitraryU with explicit generators. See also GenericArbitraryG.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryUG.

Since: 1.5.0.0

Constructors

GenericArbitraryUG

Fields

unGenericArbitraryUG :: a

newtype GenericArbitrarySingleG genList a

genericArbitrarySingle with explicit generators. See also GenericArbitraryG.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitrarySingleG.

Since: 1.5.0.0

Constructors

GenericArbitrarySingleG

Fields

unGenericArbitrarySingleG :: a

newtype GenericArbitraryRecG genList weights a

genericArbitraryRec with explicit generators. See also genericArbitraryG.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryRecG.

Since: 1.5.0.0

Constructors

GenericArbitraryRecG

Fields

unGenericArbitraryRecG :: a

newtype GenericArbitraryWith opts weights a

General generic generator with custom options.

This newtype does no shrinking. To add generic shrinking, use AndShrinking.

Uses genericArbitraryWith.

Since: 1.5.0.0

Constructors

GenericArbitraryWith

Fields

unGenericArbitraryWith :: a

newtype AndShrinking f a

Add generic shrinking to a newtype wrapper for Arbitrary, using genericShrink.

data X = ...
  deriving Arbitrary via (GenericArbitrary '[1,2,3] `AndShrinking` X)

Equivalent to:

instance Arbitrary X where
  arbitrary = genericArbitrary (1 % 2 % 3 % ())
  shrink = genericShrink

Since: 1.5.0.0

Constructors

AndShrinking a

class TypeLevelGenList a where

Since: 1.5.0.0

Associated Types

type TypeLevelGenList' a :: Type

Methods

toGenList :: Proxy a -> TypeLevelGenList' a

class TypeLevelOpts a where

Since: 1.5.0.0

Associated Types

type TypeLevelOpts' a :: Type

Methods

toOpts :: Proxy a -> TypeLevelOpts' a