Generic.Random.Internal.Generic

Description

Core implementation.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Random generators

genericArbitrary

Arguments

:: GArbitrary UnsizedOpts a

=> Weights a
List of weights for every constructor

-> Gen a

Pick a constructor with a given distribution, and fill its fields with recursive calls to arbitrary.

Example

genericArbitrary (2 % 3 % 5 % ()) :: Gen a

Picks the first constructor with probability 2/10, the second with probability 3/10, the third with probability 5/10.

genericArbitraryU :: (GArbitrary UnsizedOpts a, GUniformWeight a) => Gen a

Pick every constructor with equal probability. Equivalent to genericArbitrary uniform.

genericArbitraryU :: Gen a

genericArbitrarySingle :: (GArbitrary UnsizedOpts a, Weights_ (Rep a) ~ L c0) => Gen a

arbitrary for types with one constructor. Equivalent to genericArbitraryU, with a stricter type.

genericArbitrarySingle :: Gen a

genericArbitraryRec

Arguments

:: GArbitrary SizedOptsDef a

=> Weights a
List of weights for every constructor

-> Gen a

Decrease size at every recursive call, but don't do anything different at size 0.

genericArbitraryRec (7 % 11 % 13 % ()) :: Gen a

N.B.: This replaces the generator for fields of type [t] with listOf' arbitrary instead of listOf arbitrary (i.e., arbitrary for lists).

genericArbitraryG :: GArbitrary (SetGens genList UnsizedOpts) a => genList -> Weights a -> Gen a

genericArbitrary with explicit generators.

Example

genericArbitraryG customGens (17 % 19 % ())

where, the generators for String and Int fields are overridden as follows, for example:

customGens :: Gen String :+ Gen Int
customGens =
  (filter (/= 'NUL') <$> arbitrary) :+
  (getNonNegative <$> arbitrary)

Note on multiple matches

Multiple generators may match a given field: the first will be chosen.

genericArbitraryUG :: (GArbitrary (SetGens genList UnsizedOpts) a, GUniformWeight a) => genList -> Gen a

genericArbitraryU with explicit generators. See also genericArbitraryG.

genericArbitrarySingleG :: (GArbitrary (SetGens genList UnsizedOpts) a, Weights_ (Rep a) ~ L c0) => genList -> Gen a

genericArbitrarySingle with explicit generators. See also genericArbitraryG.

genericArbitraryRecG

Arguments

:: GArbitrary (SetGens genList SizedOpts) a

=> genList

-> Weights a
List of weights for every constructor

-> Gen a

genericArbitraryRec with explicit generators. See also genericArbitraryG.

genericArbitraryWith :: GArbitrary opts a => opts -> Weights a -> Gen a

General generic generator with custom options.

Internal

type family Weights_ (f :: Type -> Type) :: Type where ...

Equations

Weights_ (f :+: g) = Weights_ f :| Weights_ g

Weights_ (M1 D _c f) = Weights_ f

Weights_ (M1 C ('MetaCons c _i _j) _f) = L c

data a :| b

Constructors

N a Int b

data L (c :: Symbol)

Constructors

L

data Weights a

Trees of weights assigned to constructors of type a, rescaled to obtain a probability distribution.

Two ways of constructing them.

(x1 % x2 % ... % xn % ()) :: Weights a
uniform :: Weights a

Using (%), there must be exactly as many weights as there are constructors.

uniform is equivalent to (1 % ... % 1 % ()) (automatically fills out the right number of 1s).

Constructors

Weights (Weights_ (Rep a)) Int

newtype W (c :: Symbol)

Type of a single weight, tagged with the name of the associated constructor for additional compile-time checking.

((9 :: W "Leaf") % (8 :: W "Node") % ())

Constructors

W Int

weights :: (Weights_ (Rep a), Int, ()) -> Weights a

A smart constructor to specify a custom distribution. It can be omitted for the % operator is overloaded to insert it.

uniform :: UniformWeight_ (Rep a) => Weights a

Uniform distribution.

type family First a :: Symbol where ...

Equations

First (a :| _b) = First a

First (L c) = c

type family First' w where ...

Equations

First' (Weights a) = First (Weights_ (Rep a))

First' (a, Int, r) = First a

type family Prec' w where ...

Equations

Prec' (Weights a) = Prec (Weights_ (Rep a)) ()

Prec' (a, Int, r) = Prec a r

class WeightBuilder' w where

Methods

(%) :: c ~ First' w => W c -> Prec' w -> w infixr 1

A binary constructor for building up trees of weights.

class WeightBuilder a where

Associated Types

type Prec a r

Methods

(%.) :: c ~ First a => W c -> Prec a r -> (a, Int, r)

class UniformWeight a where

Methods

uniformWeight :: (a, Int)

class UniformWeight (Weights_ f) => UniformWeight_ f

class UniformWeight_ (Rep a) => GUniformWeight a

Derived uniform distribution of constructors for a.

newtype Options (c :: Coherence) (s :: Sizing) (genList :: Type)

Type-level options for GArbitrary.

Note: it is recommended to avoid referring to the Options type explicitly in code, as the set of options may change in the future. Instead, use the provided synonyms (UnsizedOpts, SizedOpts, SizedOptsDef) and the setter SetOptions (abbreviated as (<+)).

Constructors

Options

Fields

_generators :: genList

type family SetOptions (x :: k) (o :: Type) :: Type

Setter for Options.

This subsumes the other setters: SetSized, SetUnsized, SetGens.

Since: 1.4.0.0

type (<+) o x = SetOptions x o infixl 1

Infix flipped synonym for Options.

Since: 1.4.0.0

type UnsizedOpts = Options 'INCOHERENT 'Unsized ()

type SizedOpts = Options 'INCOHERENT 'Sized ()

type SizedOptsDef = Options 'INCOHERENT 'Sized (Gen1 [] :+ ())

type CohUnsizedOpts = Options 'COHERENT 'Unsized ()

Like UnsizedOpts, but using coherent instances by default.

Since: 1.4.0.0

type CohSizedOpts = Options 'COHERENT 'Sized ()

Like SizedOpts, but using coherent instances by default.

Since: 1.4.0.0

setOpts :: forall x o. Coercible o (SetOptions x o) => o -> SetOptions x o

Coerce an Options value between types with the same representation.

Since: 1.4.0.0

unsizedOpts :: UnsizedOpts

Default options for unsized generators.

sizedOpts :: SizedOpts

Default options for sized generators.

sizedOptsDef :: SizedOptsDef

Default options overriding the list generator using listOf'.

cohUnsizedOpts :: CohUnsizedOpts

Like unsizedOpts, but using coherent instances by default.

cohSizedOpts :: CohSizedOpts

Like sizedOpts but using coherent instances by default.

data Sizing

Whether to decrease the size parameter before generating fields.

The Sized option makes the size parameter decrease in the following way: - Constructors with one field decrease the size parameter by 1 to generate that field. - Constructors with more than one field split the size parameter among all fields; the size parameter is rounded down to then be divided equally.

Constructors

Sized
Decrease the size parameter when running generators for fields

Unsized
Don't touch the size parameter

type family SizingOf opts :: Sizing

type family SetSized (o :: Type) :: Type

type family SetUnsized (o :: Type) :: Type

setSized :: Options c s g -> Options c 'Sized g

setUnsized :: Options c s g -> Options c 'Unsized g

data Coherence

For custom generators to work with parameterized types, incoherent instances must be used internally. In practice, the resulting behavior is what users want 100% of the time, so you should forget this option even exists.

Details

Expand

The default configuration of generic-random does a decent job if we trust GHC implements precisely the instance resolution algorithm as described in the GHC manual:

https://downloads.haskell.org/ghc/latest/docs/html/users_guide/glasgow_exts.html#overlapping-instances

While that assumption holds in practice, it is overly context-dependent (to know the context leading to a particular choice, we must replay the whole resolution algorithm). In particular, this algorithm may find one solution, but it is not guaranteed to be unique: the behavior of the program is dependent on implementation details.

An notable property to consider of an implicit type system (such as type classes) is coherence: the behavior of the program is stable under specialization.

This sounds nice on paper, but actually leads to surprising behavior for generic implementations with parameterized types, such as generic-random.

To address that, the coherence property can be relaxd by users, by explicitly allowing some custom generators to be chosen incoherently. With appropriate precautions, it is possible to ensure a weaker property which nevertheless helps keep type inference predictable: when a solution is found, it is unique. (This is assuredly weaker, i.e., is not stable under specialization.)

Since: 1.4.0.0

Constructors

INCOHERENT
Match custom generators incoherently.

COHERENT
Match custom generators coherently by default (can be manually bypassed with Incoherent).

type family CoherenceOf (o :: Type) :: Coherence

newtype Incoherent g

Match this generator incoherently when the COHERENT option is set.

Constructors

Incoherent g

data a :+ b infixr 1

Heterogeneous list of generators.

Constructors

a :+ b infixr 1

type family GeneratorsOf opts :: Type

class HasGenerators opts where

Methods

generators :: opts -> GeneratorsOf opts

setGenerators :: genList -> Options c s g0 -> Options c s genList

Define the set of custom generators.

Note: for recursive types which can recursively appear inside lists or other containers, you may want to include a custom generator to decrease the size when generating such containers.

See also the Note about lists in Generic.Random.Tutorial.

type family SetGens (g :: Type) opts

newtype FieldGen (s :: Symbol) a

Custom generator for record fields named s.

If there is a field named s with a different type, this will result in a type error.

Constructors

FieldGen

Fields

unFieldGen :: Gen a

fieldGen :: proxy s -> Gen a -> FieldGen s a

FieldGen constructor with the field name given via a proxy.

newtype ConstrGen (c :: Symbol) (i :: Nat) a

Custom generator for the i-th field of the constructor named c. Fields are 0-indexed.

Constructors

ConstrGen

Fields

unConstrGen :: Gen a

constrGen :: proxy '(c, i) -> Gen a -> ConstrGen c i a

ConstrGen constructor with the constructor name given via a proxy.

newtype Gen1 f

Custom generators for "containers" of kind Type -> Type, parameterized by the generator for "contained elements".

A custom generator Gen1 f will be used for any field whose type has the form f x, requiring a generator of x. The generator for x will be constructed using the list of custom generators if possible, otherwise an instance Arbitrary x will be required.

Constructors

Gen1

Fields

unGen1 :: forall a. Gen a -> Gen (f a)

newtype Gen1_ f

Custom generators for unary type constructors that are not "containers", i.e., which don't require a generator of a to generate an f a.

A custom generator Gen1_ f will be used for any field whose type has the form f x.

Constructors

Gen1_

Fields

unGen1_ :: forall a. Gen (f a)

vectorOf' :: Int -> Gen a -> Gen [a]

An alternative to vectorOf that divides the size parameter by the length of the list.

listOf' :: Gen a -> Gen [a]

An alternative to listOf that divides the size parameter by the length of the list. The length follows a geometric distribution of parameter 1/(sqrt size + 1).

listOf1' :: Gen a -> Gen [a]

An alternative to listOf1 (nonempty lists) that divides the size parameter by the length of the list. The length (minus one) follows a geometric distribution of parameter 1/(sqrt size + 1).

geom :: Int -> Gen Int

Geometric distribution of parameter 1/(sqrt n + 1) (n >= 0).

class GA opts f where

Generic Arbitrary

Methods

ga :: opts -> Weights_ f -> Int -> Gen (f p)

class (Generic a, GA opts (Rep a)) => GArbitrary opts a

Generic Arbitrary

gaSum' :: GASum opts f => opts -> Weights_ f -> Int -> Gen (f p)

class GASum opts f where

Methods

gaSum :: opts -> Int -> Weights_ f -> Gen (f p)

class GAProduct (s :: Sizing) (c :: Maybe Symbol) opts f where

Methods

gaProduct :: proxys '(s, c) -> opts -> Gen (f p)

class GAProduct' (c :: Maybe Symbol) (i :: Nat) opts f where

Methods

gaProduct' :: proxy '(c, i) -> opts -> Gen (f p)

type family Arity f :: Nat where ...

Equations

Arity (f :*: g) = Arity f + Arity g

Arity (M1 _i _c _f) = 1

class FindGen (i :: AInstr) (s :: AStore) (g :: Type) (gs :: Type) (a :: Type) where

Given a list of custom generators g :+ gs, find one that applies, or use Arbitrary a by default.

g and gs follow this little state machine:

          g,      gs | result
---------------------+-----------------------------
         (),      () | END
         (), g :+ gs | g, gs
         (),      g  | g, () when g is not (_ :+ _)
     g :+ h,      gs | g, h :+ gs
      Gen a,      gs | END if g matches, else ((), gs)
 FieldGen a,      gs | idem
ConstrGen a,      gs | idem
     Gen1 a,      gs | idem
    Gen1_ a,      gs | idem

Methods

findGen :: (Proxy i, Proxy s, FullGenListOf s) -> g -> gs -> Gen a

data AInstr

Constructors

Shift

Match Coherence

MatchCoh Bool

data AStore

Constructors

S Type Coherence ASel

type ASel = (Maybe Symbol, Nat, Maybe Symbol)

iShift :: Proxy 'Shift

type family FullGenListOf (s :: AStore) :: Type where ...

Equations

FullGenListOf ('S fg _coh _sel) = fg

type family ACoherenceOf (s :: AStore) :: Coherence where ...

Equations

ACoherenceOf ('S _fg coh _sel) = coh

type family ASelOf (s :: AStore) :: ASel where ...

Equations

ASelOf ('S _fg _coh sel) = sel

type DummySel = '('Nothing, 0, 'Nothing)

type family Name (d :: Meta) :: Maybe Symbol

Get the name contained in a Meta tag.

type family Matches (s :: ASel) (g :: Type) (a :: Type) :: Bool where ...

Equations

Matches _sel (Gen b) a = b == a

Matches _sel (Gen1_ f) (f a) = 'True

Matches _sel (Gen1_ f) a = 'False

Matches _sel (Gen1 f) (f a) = 'True

Matches _sel (Gen1 f) a = 'False

Matches '(_c, i, s) (FieldGen s1 b) a = (s == 'Just s1) && (b == a)

Matches '(c, i, _s) (ConstrGen c1 j b) a = (c == 'Just c1) && ((i == j) && (b == a))

newtype Weighted a

Constructors

Weighted (Maybe (Int -> Gen a, Int))

liftGen :: Gen a -> Weighted a